+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog #
 
+## 1.0.2 - Nov 9, 2015
+* Perf improvements and internal cleanup.
+* Remove need for babel-polyfill dep.
+* Allow generators to be directly as iterators.
+** `Array.from(pep.seq(1, 2, 3)) === ['1', '2', '3']
+
 ## 1.0.1 - Nov 8, 2015
 * Corrected babel-polyfill being listed as a dev dep.